% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_prior_grid}
\alias{create_prior_grid}
\title{Create a prior grid}
\usage{
create_prior_grid(prior_input)
}
\arguments{
\item{prior_input}{A numeric matrix}
}
\value{
A (normalized) numeric matrix
}
\description{
'create_prior_grid' returns a valid prior from a given numeric matrix
}
\details{
This is a function that takes any numeric matrix and returns another numeric matrix which is a valid probability
density function that has a "prior" attribute assigned to it. The input doesn't have to be a valid PDF the function will
do this so when using this package its always good to use this function before using any of the search functions.
}
\examples{
create_prior_grid(matrix(c(1,2,3,4),nrow = 1, ncol = 4))


}
