% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_matrix.R
\name{generate_sample_matrix}
\alias{generate_sample_matrix}
\title{Generate sample matrices}
\usage{
generate_sample_matrix(n)
}
\arguments{
\item{n}{An integer representing the number of rows and columns for the square matrices}
}
\value{
3 x Numeric matrix
}
\description{
Generates a set of matrices for use in the search functions.
}
\details{
This is a helper function that generates 3 matrices to be used in the search functions in this package.
The function will assign 3 numeric functions to the global environment: "prior_grid", "pod_grid" and "cost_grid".
"prior_grid" is a normalized matrix of value generated by the \link[MASS]{kde2d} function from the MASS package.
Both "pod_grid" and "cost_grid" are random values sampled from the Uniform(0,1) distribution.
}
\examples{
generate_sample_matrix(5)

}
