% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimal_search_sequence_posteriors}
\alias{optimal_search_sequence_posteriors}
\title{Optimal search posteriors}
\usage{
optimal_search_sequence_posteriors(
  prior_grid,
  pod_grid,
  cost_grid,
  num_searches
)
}
\arguments{
\item{prior_grid}{A numeric matrix that is a valid probability density function. Use \code{\link{create_prior_grid}} for validation.}

\item{pod_grid}{A numeric matrix. All matrix entries must be greater than 0 and less than 1.}

\item{cost_grid}{A numeric matrix representing the cost of searching each grid square}
}
\value{
List of the same length as the search sequence. Each list element contains a posterior (Numeric Matrix), overall probability
of detection (Double) and the grid square searched (String)
}
\description{
Calculates the posteriors of the optimal search of a given length
}
\details{
This function takes a cost grid in addition to the standard prior grid and probability of detection grid and will calculate the optimal
search sequence for a given length.
}
\examples{

generate_sample_matrix(5)
optimal_search_sequence_posteriors(prior_grid, pod_grid, cost_grid, 10)

}
