% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posterior_sequence_long}
\alias{posterior_sequence_long}
\title{Melt Search sequence output}
\usage{
posterior_sequence_long(search_posteriors)
}
\arguments{
\item{search_posteriors, }{the list output of a search function}
}
\value{
A dataframe
}
\description{
Melt the output of search sequence functions into long data-frame format
}
\details{
This function melts the output of a search function to a long data-frame format to allow for easier analysis
of the searcg
}
\examples{
generate_sample_matrix(5)
search_sequence <- matrix(c(1,5,2,5,3,5,4,5,5,5), nrow = 5, ncol = 2)
search_posteriors <- search_sequence_posteriors(prior_grid, pod_grid, search_sequence)
posterior_sequence_long(search_posteriors)

}
