% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_posterior_neg}
\alias{update_posterior_neg}
\title{Calculate the posterior distribution}
\usage{
update_posterior_neg(prior_grid, pod_grid, row, col)
}
\arguments{
\item{prior_grid}{A numeric matrix that is a valid probability density function. Use \code{\link{create_prior_grid}} for validation.}

\item{pod_grid}{A numeric matrix. All matrix entries must be greater than 0 and less than 1.}

\item{row}{Integer representing the row index}

\item{col}{Integer representing the column index}
}
\value{
Numeric matrix
}
\description{
Calculate the posterior distribution after a negative search of a given grid square
}
\details{
This function will calculate the posterior distribution from a given prior matrix, probability of detection matrix
and a column and row index.
}
\examples{
generate_sample_matrix(5)
update_posterior_neg(prior_grid, pod_grid, 1, 1)

}
