% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_matrix.R
\name{generate_sample_matrix}
\alias{generate_sample_matrix}
\title{Generate sample matrices}
\usage{
generate_sample_matrix(n)
}
\arguments{
\item{n}{An integer representing the number of rows and columns for the square matrices}
}
\value{
A Prior, Probability of Detection (POD) and a cost matrix stored in the global environment
}
\description{
Generate sample matrices
}
